+++ packages/insomnia/src/ui/components/dropdowns/workspace-sync-dropdown.tsx
@@ -4,7 +4,7 @@ import { useRouteLoaderData } from 'react-router';
 import { isGitProject, isRemoteProject } from '../../../models/project';
 import { useOrganizationPermissions } from '../../hooks/use-organization-features';
 import { useRootLoaderData } from '../../routes/root';
-import type { WorkspaceLoaderData } from '../../routes/workspace';
+import type { WorkspaceLoaderData } from '../../routes/$organizationId.project.$projectId.workspace';
 import { GitProjectSyncDropdown } from './git-project-sync-dropdown';
 import { GitSyncDropdown } from './git-sync-dropdown';
 import { LocalProjectBar } from './local-project-bar';
