+++ packages/insomnia/src/ui/components/settings/import-export.tsx
@@ -23,11 +23,11 @@ import { AskModal } from 'insomnia/src/ui/components/modals/ask-modal';
 import { ExportRequestsModal } from 'insomnia/src/ui/components/modals/export-requests-modal';
 import { ImportModal } from 'insomnia/src/ui/components/modals/import-modal';
 import { SelectModal } from 'insomnia/src/ui/components/modals/select-modal';
+import type { WorkspaceLoaderData } from 'insomnia/src/ui/routes/$organizationId.project.$projectId.workspace';
 import { useOrganizationLoaderData } from 'insomnia/src/ui/routes/organization';
 import type { ListWorkspacesLoaderData } from 'insomnia/src/ui/routes/project';
 import { useRootLoaderData } from 'insomnia/src/ui/routes/root';
 import type { UntrackedProjectsLoaderData } from 'insomnia/src/ui/routes/untracked-projects';
-import type { WorkspaceLoaderData } from 'insomnia/src/ui/routes/workspace';
 import React, { type FC, Fragment, useEffect, useState } from 'react';
 import { Button, Heading, ListBox, ListBoxItem, Popover, Select, SelectValue } from 'react-aria-components';
 import { useFetcher, useParams } from 'react-router';
