+++ packages/insomnia/src/ui/context/app/insomnia-event-stream-context.tsx
@@ -9,7 +9,7 @@ import { insomniaFetch } from '../../../ui/insomniaFetch';
 import { avatarImageCache } from '../../hooks/image-cache';
 import type { ProjectIdLoaderData } from '../../routes/project';
 import { useRootLoaderData } from '../../routes/root';
-import type { WorkspaceLoaderData } from '../../routes/workspace';
+import type { WorkspaceLoaderData } from '../../routes/$organizationId.project.$projectId.workspace';
 
 const InsomniaEventStreamContext = createContext<{
   presence: UserPresence[];
