+++ packages/insomnia/src/ui/context/nunjucks/use-nunjucks.ts
@@ -6,7 +6,7 @@ import { NUNJUCKS_TEMPLATE_GLOBAL_PROPERTY_NAME } from '../../../templating';
 import type { HandleRender, RenderContextOptions } from '../../../templating/types';
 import { getKeys } from '../../../templating/utils';
 import type { RequestLoaderData } from '../../routes/$organizationId.project.$projectId.workspace.$workspaceId.debug.request.$requestId';
-import type { WorkspaceLoaderData } from '../../routes/workspace';
+import type { WorkspaceLoaderData } from '../../routes/$organizationId.project.$projectId.workspace';
 let getRenderContextPromiseCache: any = {};
 
 export interface UseNunjucksOptions {
