+++ packages/insomnia/src/ui/hooks/use-document-title.ts
@@ -3,7 +3,7 @@ import { useRouteLoaderData } from 'react-router';
 
 import { getProductName } from '../../common/constants';
 import type { RequestLoaderData } from '../routes/$organizationId.project.$projectId.workspace.$workspaceId.debug.request.$requestId';
-import type { WorkspaceLoaderData } from '../routes/workspace';
+import type { WorkspaceLoaderData } from '../routes/$organizationId.project.$projectId.workspace';
 export const useDocumentTitle = () => {
   const { activeWorkspace, activeEnvironment, activeProject } = useRouteLoaderData(
     ':workspaceId',
