+++ packages/insomnia/src/ui/hooks/use-vcs-version.ts
@@ -5,7 +5,7 @@ import { useParams } from 'react-router';
 import { type ChangeBufferEvent, database } from '../../common/database';
 import type { BaseModel } from '../../models';
 import type { ProjectLoaderData } from '../routes/project';
-import type { WorkspaceLoaderData } from '../routes/workspace';
+import type { WorkspaceLoaderData } from '../routes/$organizationId.project.$projectId.workspace';
 // We use this hook to determine if the active request has been updated from the system (not the user typing)
 // For example, by pulling a new version from the remote, switching branches, etc.
 export function useActiveRequestSyncVCSVersion() {
